+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Bug Fixes
 
 -   Constructor parameters which share a name with a property on a parent class will no longer inherit the comment on the parent class, #2636.
+-   Packages mode will now attempt to use the comment declared in the comment class for inherited members, #2622.
 
 ## v0.26.4 (2024-07-10)