+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 -   Constructor parameters which share a name with a property on a parent class will no longer inherit the comment on the parent class, #2636.
 -   Packages mode will now attempt to use the comment declared in the comment class for inherited members, #2622.
+-   TypeDoc no longer crashes when `@document` includes an empty file, #2638.
+
+### Thanks!
+
+-   @bladerunner2020
 
 ## v0.26.4 (2024-07-10)